requirements.txt
asgiref==3.7.2
boto3==1.29.0
botocore==1.32.0
certifi==2023.7.22
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docutils @ file:///private/tmp/docutils-20230904-78923-189veup/docutils-0.20.1
environment==1.0.0
flit_core @ file:///private/tmp/python-flit-core-20230904-11014-fg5gk4/flit_core-3.9.0
gunicorn==21.2.0
jmespath==1.0.1
packaging @ file:///private/tmp/python-packaging-20230904-11090-7eg0k3/packaging-23.1
psycopg2==2.9.9
Pygments @ file:///private/tmp/pygments-20230904-97052-1pr7lwk/Pygments-2.16.1
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
